+++ src/mako/i18n/I18n.php
@@ -30,14 +30,14 @@ class I18n
 	 *
 	 * @var string
 	 */
-	public const PLURALIZATION_TAG_REGEX = '/\<pluralize:([0-9]+)\>(\w*)\<\/pluralize\>/iu';
+	protected const PLURALIZATION_TAG_REGEX = '/\<pluralize:([0-9]+)\>(\w*)\<\/pluralize\>/iu';
 
 	/**
 	 * Regex that matches number tags.
 	 *
 	 * @var string
 	 */
-	public const NUMBER_TAG_REGEX = '/\<number(:([0-9]+)(,(.)(,(.))?)?)?\>([0-9-.e]*)\<\/number\>/iu';
+	protected const NUMBER_TAG_REGEX = '/\<number(:([0-9]+)(,(.)(,(.))?)?)?\>([0-9-.e]*)\<\/number\>/iu';
 
 	/**
 	 * Loaded language strings.